feat(dev): add SCC and jc for code analysis and JSON parsing
Install SCC (Source Code Counter) via Homebrew to enable counting lines of
code in the project.

Also install jc, a CLI tool and python library that converts the output of
popular command-line tools and file-types to JSON for easier parsing in
scripts. It is intended to be used with `/usr/bin/time`.

Brew itself is now installed in the devcontainer to support installing these
additional developer utilities.