feat: add computation of GNATcheck overhead to benchmark script
The benchmark script is extended to measure the overhead incurred by
GNATcheck itself. This is achieved by introducing a new rule that do nothing, and running GNATcheck with this
 rule for a number of iterations with different parallelization settings.

Computing the overhead allows for more accurate assessment of GNATcheck's actual performance by
distinguishing the time spent in setup and teardown from the time spent in the actual analysis.
This change will contribute to the precision and reliability of the benchmark results.