feat(benchmark): add overhead computation for AdaControl
The benchmark script now computes the overhead for AdaControl in addition
to GNATcheck. This is done by running the Adactl_benchmark.sh script with
a rule file that does no actual checks.

The results will provide a baseline for the actual analysis time of
AdaControl itself, which can be subtracted from the total analysis time to
determine the time spent in the user-defined rules.