feat(utils): add `generate-scc-metrics` command to generate code metrics
Introduces a new command `generate-scc-metrics` which uses the SCC (Source
Code Counter) tool to generate code metrics such as lines of code,
complexity, and number of files for a given Ada project.

The command takes a `project.units_by_path` file as input, which lists the
source files of the project. It runs SCC on these files and outputs the
resulting metrics to a JSON file named `<project>_scc-metrics.json`.

This will enable collecting and tracking valuable code metrics for Ada
projects to gain insights into their size, complexity and structure over
time.