refactor(runBenchmarkCommandFactory): Change 'gprPath' to 'alireTomlPath' in benchmark scripts
The variables passed to the 'clean' function in the 'process_project' and
'clean' functions have been renamed from 'gprPath' to 'alireTomlPath' in the
Adactl_benchmark.sh, gnatcheck_benchmark.sh, and runBenchmarkCommandFactory.ts
files.

This change ensures consistency and clarity in the naming of variables across
the benchmark scripts, making it evident that the path being referred to is
for the alire.toml file rather than a GPR project file.