feat(runBenchmarkCommandFactory): add support for %UNITS_BY_FILENAME%
This allows the factory to replace occurrences of "%UNITS_BY_FILENAME%"
in the command with the path to the ".units_by_filename" file
corresponding to the project's ".gpr" file.

The change enhances flexibility by enabling the command to reference
either the ".units" or ".units_by_filename" file associated with each
project being benchmarked.